feat: add faketrain for testing
The current version of the software has a fake train that consumes
events and publishes them to the topics specified in the README.md spec.
The webserver is able to send the train to the raucherecke station, and
to display status messages in a table.

The real improvements happen under the hood though: I wrote down all zod
schemas in schemas.ts and try to use them everywhere. The code feels
good and sturdy from that perspective. I'd like to soon add linting,
having a homogenous code style would help.

In the mid-term, I want to see automated testing and a proper state
machine in useTrainController that figures out how to properly control
the train depending on where it is in the grid and where the user wants
it to go.